+++ src/modules/edc-demo/components/asset-editor-dialog/data-address-type.ts
@@ -1,1 +1,1 @@
-export type DataAddressType = 'Json' | 'Rest-Api';
+export type DataAddressType = 'Custom-Datasource-Json' | 'Http';