+++ src/modules/edc-demo/components/asset-editor-dialog/model/asset-metadata-form-model.ts
@@ -1,0 +1,15 @@
+import {FormControl} from '@angular/forms';
+import {LanguageSelectItem} from '../../language-select/language-select-item';
+
+/**
+ * Form Model for AssetEditorDialog > Metadata
+ */
+export interface AssetMetadataFormModel {
+  id: FormControl<string>;
+  name: FormControl<string>;
+  version: FormControl<string>;
+  contentType: FormControl<string>;
+  description: FormControl<string>;
+  keywords: FormControl<string[]>;
+  language: FormControl<LanguageSelectItem | null>;
+}