+++ src/modules/edc-demo/components/asset-editor-dialog/model/http-datasource-header-form-model.ts
@@ -1,0 +1,15 @@
+import {FormControl, ɵFormGroupValue} from '@angular/forms';
+
+/**
+ * Form Model for AssetEditorDialog > Datasource > HTTP/REST > Header
+ */
+export interface HttpDatasourceHeaderFormModel {
+  headerName: FormControl<string>;
+  headerValue: FormControl<string>;
+}
+
+/**
+ * Form Value for AssetEditorDialog > Datasource > HTTP/REST > Header
+ */
+export type HttpDatasourceHeaderFormValue =
+  ɵFormGroupValue<HttpDatasourceHeaderFormModel>;